Rework commit "All pattern recognizers will ignore incomplete candlestick"
This solution of this problem is more complicated than I originally
estimated.

In commit, I tried to minimize the number of duplicated bull_flag
and shooting star signal, which has noticable impact on the system's
performance. Additionally, shooting star detection won't be fully
accurate if we use an incomplete candlestick.